requirements.txt
Flask==3.0.3
werkzeug==3.0.3
PyMuPDF==1.23.26
transformers==4.38.2
torch==2.2.1
prometheus-client==0.19.0
langdetect==1.0.9
gunicorn==21.2.0
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
scikit-learn==1.3.2
joblib==1.3.2
PyPDF2==3.0.1
python-docx==1.0.1
python-pptx==0.6.21
openpyxl==3.1.2
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.1
python-jose==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-dotenv==1.0.0